requirements.txt
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.3
attrs==19.2.0
awscli==1.16.296
backcall==0.1.0
black==19.3b0
boto3==1.10.32
botocore==1.13.32
cli-helpers==1.2.1
Click==7.0
colorama==0.4.1
configobj==5.0.6
decorator==4.4.0
Django==3.0
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==20.0.4
humanize==0.5.1
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
jmespath==0.9.4
parso==0.5.1
pexpect==4.7.0
pgcli==2.1.1
pgspecial==1.11.7
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.9
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
rsa==3.4.2
s3transfer==0.2.1
setproctitle==1.1.10
six==1.12.0
sqlparse==0.3.0
tabulate==0.8.5
terminaltables==3.1.0
toml==0.10.0
traitlets==4.3.2
urllib3==1.25.7
wcwidth==0.1.7
